refactor(akashi): split background loops, integrity sweeps and adapters out of akashi.go
Motion only; no function body changes.

The module root is itself package akashi, and it is where a contributor lands
first: cmd/akashi/main.go is a 59-line shim, so the real entrypoint was a
1,747-line file whose package doc advertises an embedding API while roughly
90% of its contents are background goroutines and integrity sweeps. The rest of
the package is tiny — interfaces.go 60 lines, options.go 100, types.go 96 — so
akashi.go carried all of it alone.

  - background.go: the loop scaffolding (runLoop, safeRun) and every ticker
    loop — conflict backfill and refresh, idempotency and hook-check cleanup,
    retention, claim-embedding retry, percentile refresh, auto-resolve, stale
    supersedes pruning — plus the two timeout helpers they share.
  - integrity_audit.go: integrityProofLoop through persistViolation, a
    contiguous, self-contained cluster that was wedged in the middle of the
    background-loop run for no reason but declaration order, plus
    buildIntegrityProofs.
  - adapters.go: the boundary shims (decisionHookAdapter, externalScorerAdapter,
    searcherAdapter, authHelperImpl) and the four constructor helpers that
    select an embedding provider, a conflict validator and a claim extractor.

What deliberately does NOT move: the package doc, App, New, Run, Shutdown, and
the conversion helpers toPublicDecision / toPublicConflict / derefOr.
akashi.go:16-19 documents those helpers as living there "because this is the
only file that sees both sides of the boundary", and moving them would make the
package doc false — the same drift class commit 28a564e existed to repair. This
is stated here so the next person does not read the split as unfinished and
"complete" it. akashi.go itself does not move either: the private enterprise
repo imports this package path.

Verified as pure motion rather than asserted. Hand-counted sed ranges are how a
motion PR silently clips a doc comment or drops a brace, so the blocks were cut
by a parser that attaches each declaration's comment block and walks to its
column-0 closing brace, and every destination file was re-parsed afterwards and
byte-compared against what was extracted: 19, 8 and 14 declarations, all
identical. The root package's tests needed no edits — backfill_test.go
exercises drainConflictBackfill, now declared in background.go, same package —
and `go test -list` returns the same 7 names.

One import could not be inferred: goimports does not restore the otelmetric
alias on go.opentelemetry.io/otel/metric, so integrity_audit.go failed to
compile until it was added by hand. Caught by the build, not by review.

Build, vet, golangci-lint and the lite build are all clean.
