+++ akashi.go
@@ -21,22 +21,16 @@ package akashi
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"log/slog"
-	"math/rand/v2"
 	"net/http"
 	"os"
-	"runtime/debug"
 	"sync"
 	"sync/atomic"
 	"time"
 
-	"github.com/google/uuid"
 	"github.com/joho/godotenv"
-	"go.opentelemetry.io/otel/attribute"
 	otelmetric "go.opentelemetry.io/otel/metric"
 
 	"github.com/ashita-ai/akashi/api"
@@ -44,7 +38,6 @@ import (
 	"github.com/ashita-ai/akashi/internal/authz"
 	"github.com/ashita-ai/akashi/internal/config"
 	"github.com/ashita-ai/akashi/internal/conflicts"
-	"github.com/ashita-ai/akashi/internal/integrity"
 	"github.com/ashita-ai/akashi/internal/mcp"
 	"github.com/ashita-ai/akashi/internal/model"
 	"github.com/ashita-ai/akashi/internal/ratelimit"
@@ -700,819 +693,9 @@ func (a *App) Shutdown(ctx context.Context) error {
 
 // ── Background loop helper ─────────────────────────────────────────────────────
 
-// runLoop runs fn on every tick of interval until ctx is cancelled.
-// It recovers from panics in fn and logs them so a single bad tick
-// cannot kill a background goroutine.
-func (a *App) runLoop(ctx context.Context, name string, interval time.Duration, fn func(ctx context.Context)) {
-	ticker := time.NewTicker(interval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-			a.safeRun(ctx, name, fn)
-		}
-	}
-}
-
-// safeRun executes fn with panic recovery. Extracted from runLoop so that a
-// one-shot pass performed outside the ticker (such as the conflict backfill's
-// startup sweep) carries the same protection a ticked pass does — otherwise a
-// panic on the startup path takes the process down where the identical panic on
-// the periodic path would only be logged.
-func (a *App) safeRun(ctx context.Context, name string, fn func(ctx context.Context)) {
-	defer func() {
-		if r := recover(); r != nil {
-			a.logger.Error("panic in background loop",
-				"loop", name,
-				"panic", r,
-				"stack", string(debug.Stack()),
-			)
-		}
-	}()
-	fn(ctx)
-}
-
 // ── Background loops (moved from cmd/akashi/main.go) ──────────────────────────
 
-func (a *App) conflictBackfillLoop(ctx context.Context) {
-	// Warm up the Ollama model before the backfill starts. Without this the
-	// first backfill request pays the full cold-start cost (model load from
-	// disk), which can exceed the per-call timeout on CPU hardware.
-	if v, ok := a.conflictScorer.Validator().(*conflicts.OllamaValidator); ok {
-		a.logger.Info("conflict backfill: warming up ollama model")
-		if err := v.Warmup(ctx); err != nil {
-			a.logger.Warn("conflict backfill: ollama warmup failed (will proceed anyway)", "error", err)
-		} else {
-			a.logger.Info("conflict backfill: ollama model ready")
-		}
-	}
-	// Drain whatever is already pending, then keep sweeping. This ran exactly
-	// once with a hardcoded batch of 500 until #745: the function was registered
-	// alongside the genuine ticker loops and named like one, but scored a single
-	// batch and returned. Any backlog larger than one batch — an operator-forced
-	// rescore being the obvious case, which resets every decision's scored mark —
-	// silently stopped at 500 and never resumed without a process restart.
-	a.safeRun(ctx, "conflictBackfill", a.sweepConflictBackfill)
-	a.runLoop(ctx, "conflictBackfill", a.cfg.ConflictBackfillInterval, a.sweepConflictBackfill)
-}
-
-// conflictBackfillMaxBatches bounds a single backfill sweep. BackfillScoring
-// marks each decision scored as it goes, so the drain normally ends when a short
-// batch comes back; this bound only matters if a decision is somehow never
-// marked, which would otherwise re-fetch the same rows forever and spin the LLM
-// judge hot. It is deliberately high enough that a real backlog drains in one
-// sweep (200 * the default 500 = 100k decisions).
-const conflictBackfillMaxBatches = 200
-
-// sweepConflictBackfill drains the pending-scoring queue and reports what it did.
-func (a *App) sweepConflictBackfill(ctx context.Context) {
-	n := drainConflictBackfill(ctx, a.conflictScorer, a.cfg.ConflictBackfillBatchSize,
-		conflictBackfillMaxBatches, a.logger)
-	if n > 0 {
-		a.logger.Info("conflict scoring backfill complete", "decisions_scored", n)
-	}
-}
-
-// backfillScorer is the slice of the conflict scorer the drain needs, narrowed to
-// an interface so the drain logic is testable without a database or an LLM.
-type backfillScorer interface {
-	BackfillScoring(ctx context.Context, batchSize int) (int, error)
-}
-
-// drainConflictBackfill scores pending decisions batch by batch until the queue
-// is empty, returning the total number of decisions processed.
-//
-// Termination rests on BackfillScoring's contract: it processes at most
-// batchSize decisions and marks each one scored, so a batch shorter than
-// batchSize means nothing is left to score. TestBackfillScoring_MarksDecisionsScored
-// pins that invariant — a second call over the same data returns 0.
-//
-// A batch error stops the sweep rather than retrying in place: the next tick
-// retries from the same position, and continuing past an error risks hammering a
-// failing database or judge. The error is logged with the count achieved so far
-// so a partial sweep is never reported as a clean one.
-func drainConflictBackfill(ctx context.Context, s backfillScorer, batchSize, maxBatches int, logger *slog.Logger) int {
-	total := 0
-	for range maxBatches {
-		if ctx.Err() != nil {
-			return total
-		}
-		n, err := s.BackfillScoring(ctx, batchSize)
-		total += n
-		if err != nil {
-			logger.Warn("conflict scoring backfill failed",
-				"error", err, "decisions_scored", total)
-			return total
-		}
-		if n < batchSize {
-			return total
-		}
-	}
-	logger.Warn("conflict scoring backfill hit its per-sweep batch bound — remaining decisions wait for the next sweep",
-		"max_batches", maxBatches, "batch_size", batchSize, "decisions_scored", total)
-	return total
-}
-
-func (a *App) conflictRefreshLoop(ctx context.Context) {
-	ticker := time.NewTicker(a.cfg.ConflictRefreshInterval)
-	defer ticker.Stop()
-
-	lastNotifiedAt := make(map[uuid.UUID]time.Time)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-			opCtx, cancel := context.WithTimeout(ctx, conflictRefreshTimeout(a.cfg.ConflictRefreshInterval))
-			if err := a.db.RefreshConflicts(opCtx); err != nil {
-				cancel()
-				a.logger.Warn("conflict refresh failed", "error", err)
-				continue
-			}
-			if err := a.db.RefreshAgentState(opCtx); err != nil {
-				a.logger.Warn("agent state refresh failed", "error", err)
-			}
-
-			orgIDs, err := a.db.ListOrganizationIDs(opCtx)
-			if err != nil {
-				cancel()
-				a.logger.Warn("conflict org list failed", "error", err)
-				continue
-			}
-
-			var totalNotified int
-			for _, orgID := range orgIDs {
-				since, ok := lastNotifiedAt[orgID]
-				if !ok {
-					since = time.Now().UTC()
-					lastNotifiedAt[orgID] = since
-				}
-
-				newConflicts, err := a.db.NewConflictsSinceByOrg(opCtx, orgID, since, 1000)
-				if err != nil {
-					a.logger.Warn("new conflicts query failed", "error", err, "org_id", orgID)
-					continue
-				}
-
-				for _, c := range newConflicts {
-					// Fire OnConflictDetected hooks asynchronously.
-					if len(a.decisionHooks) > 0 {
-						conflict := c
-						hooks := a.decisionHooks
-						logger := a.logger
-						go func() {
-							hookCtx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-							defer cancel()
-							for _, h := range hooks {
-								if err := h.OnConflictDetected(hookCtx, conflict); err != nil {
-									logger.Warn("event hook OnConflictDetected failed", "error", err)
-								}
-							}
-						}()
-					}
-
-					payload, err := json.Marshal(map[string]any{
-						"org_id":        c.OrgID,
-						"conflict_kind": c.ConflictKind,
-						"decision_a_id": c.DecisionAID,
-						"decision_b_id": c.DecisionBID,
-						"agent_a":       c.AgentA,
-						"agent_b":       c.AgentB,
-						"decision_type": c.DecisionType,
-					})
-					if err != nil {
-						a.logger.Warn("conflict notify marshal failed", "error", err)
-						continue
-					}
-					if err := a.db.Notify(opCtx, storage.ChannelConflicts, string(payload)); err != nil {
-						a.logger.Warn("conflict notify failed", "error", err)
-					}
-					if c.DetectedAt.After(lastNotifiedAt[orgID]) {
-						lastNotifiedAt[orgID] = c.DetectedAt
-					}
-					totalNotified++
-				}
-			}
-			cancel()
-
-			if totalNotified > 0 {
-				a.logger.Info("conflict notifications sent", "count", totalNotified)
-			}
-		}
-	}
-}
-
-func (a *App) integrityProofLoop(ctx context.Context) {
-	a.runLoop(ctx, "integrityProof", a.cfg.IntegrityProofInterval, func(ctx context.Context) {
-		opCtx, cancel := context.WithTimeout(ctx, 5*time.Minute)
-		defer cancel()
-		buildIntegrityProofs(opCtx, a.db, a.logger)
-		if hasNull, err := a.db.HasDecisionsWithNullSearchVector(opCtx); err == nil && hasNull {
-			a.logger.Warn("decisions with NULL search_vector detected — FTS excludes these rows; check trigger and migration 022 backfill")
-		}
-	})
-}
-
-func (a *App) integrityAuditLoop(ctx context.Context) {
-	// Jitter the first tick so multiple replicas don't audit at the same wall-clock time.
-	jitter := time.Duration(rand.IntN(int(a.cfg.IntegrityAuditInterval))) //nolint:gosec // jitter, not security
-	select {
-	case <-ctx.Done():
-		return
-	case <-time.After(jitter):
-	}
-
-	ticker := time.NewTicker(a.cfg.IntegrityAuditInterval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-			opCtx, cancel := context.WithTimeout(ctx, a.cfg.IntegrityAuditTimeout)
-			a.auditIntegrityProofs(opCtx)
-			cancel()
-		}
-	}
-}
-
-// integrityFullAuditLoop runs an exhaustive integrity audit across all orgs
-// at a lower frequency (default: every 24h). Unlike the sampling audit, this
-// checks IntegrityFullAuditProofs proofs per org for every org.
-// Disabled when IntegrityFullAuditInterval is 0.
-func (a *App) integrityFullAuditLoop(ctx context.Context) {
-	if a.cfg.IntegrityFullAuditInterval <= 0 {
-		return
-	}
-
-	// Jitter the first tick so multiple replicas don't all run the full sweep simultaneously.
-	jitter := time.Duration(rand.IntN(int(a.cfg.IntegrityFullAuditInterval))) //nolint:gosec // jitter, not security
-	select {
-	case <-ctx.Done():
-		return
-	case <-time.After(jitter):
-	}
-
-	ticker := time.NewTicker(a.cfg.IntegrityFullAuditInterval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-			opCtx, cancel := context.WithTimeout(ctx, a.cfg.IntegrityAuditTimeout)
-			a.fullIntegrityAudit(opCtx)
-			cancel()
-		}
-	}
-}
-
-// auditIntegrityProofs picks one org per tick (sampling-based, not exhaustive)
-// and verifies its most recent Merkle proofs by:
-// (1) recomputing each root from decision content hashes, and
-// (2) checking chain linkage via previous_root.
-// Coverage is probabilistic: with N orgs, each org is audited roughly every
-// N * IntegrityAuditInterval. Only the 10 newest proofs per org are checked.
-// All results (pass and fail) are persisted to integrity_audit_results.
-func (a *App) auditIntegrityProofs(ctx context.Context) {
-	// Use offset-based selection instead of loading the full org table.
-	// CountOrganizations is a cheap count(*) on the PK index.
-	orgCount, err := a.db.CountOrganizations(ctx)
-	if err != nil {
-		a.logger.Warn("integrity audit: count orgs failed", "error", err)
-		return
-	}
-	if orgCount == 0 {
-		return
-	}
-
-	// Round-robin through orgs using atomic counter + offset.
-	idx := a.auditOrgCounter.Add(1) - 1
-	offset := int(idx % uint64(orgCount)) //nolint:gosec // orgCount is validated positive above; modulo result fits int
-	orgID, err := a.db.GetOrgIDByOffset(ctx, offset)
-	if err != nil {
-		a.logger.Warn("integrity audit: get org by offset failed", "offset", offset, "error", err)
-		return
-	}
-	if orgID == uuid.Nil {
-		// Org count changed between count and offset query — skip this tick.
-		return
-	}
-
-	proofs, err := a.db.GetRecentIntegrityProofs(ctx, orgID, 10)
-	if err != nil {
-		a.logger.Warn("integrity audit: failed to fetch proofs", "org_id", orgID, "error", err)
-		return
-	}
-	if len(proofs) == 0 {
-		return
-	}
-
-	results := a.verifyProofsForOrg(ctx, orgID, proofs, "sample")
-
-	// Persist results with a dedicated context so that a timeout during
-	// verification doesn't cause us to lose the audit record. An audit that
-	// detected tampering but failed to persist it is worse than one that
-	// never ran — it creates a false sense of coverage.
-	persistCtx, persistCancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer persistCancel()
-	if err := a.db.InsertIntegrityAuditResults(persistCtx, results); err != nil {
-		a.logger.Error("integrity audit: FAILED TO PERSIST results — audit ran but paper trail is missing",
-			"org_id", orgID, "proofs_checked", len(proofs), "error", err)
-	}
-
-	a.logger.Info("integrity audit completed", "org_id", orgID, "proofs_checked", len(proofs))
-}
-
-// fullIntegrityAudit checks every org exhaustively using IntegrityFullAuditProofs
-// proofs per org. All results are persisted. Each org gets its own timeout
-// (IntegrityAuditTimeout) so that slow orgs don't starve later ones.
-func (a *App) fullIntegrityAudit(ctx context.Context) {
-	orgIDs, err := a.db.ListOrganizationIDs(ctx)
-	if err != nil {
-		a.logger.Warn("integrity full audit: list orgs failed", "error", err)
-		return
-	}
-
-	var totalProofs, totalFailures int
-	for _, orgID := range orgIDs {
-		if ctx.Err() != nil {
-			a.logger.Warn("integrity full audit: cancelled before completing all orgs",
-				"completed", totalProofs, "remaining_orgs", len(orgIDs))
-			return
-		}
-
-		// Per-org timeout prevents one large org from consuming the entire sweep budget.
-		orgCtx, orgCancel := context.WithTimeout(ctx, a.cfg.IntegrityAuditTimeout)
-
-		proofs, err := a.db.GetRecentIntegrityProofs(orgCtx, orgID, a.cfg.IntegrityFullAuditProofs)
-		if err != nil {
-			a.logger.Warn("integrity full audit: failed to fetch proofs",
-				"org_id", orgID, "error", err)
-			orgCancel()
-			continue
-		}
-		if len(proofs) == 0 {
-			orgCancel()
-			continue
-		}
-
-		results := a.verifyProofsForOrg(orgCtx, orgID, proofs, "full")
-		orgCancel()
-
-		// Persist with a dedicated context — same rationale as the sampling audit:
-		// verification results must survive even if the verification context expired.
-		persistCtx, persistCancel := context.WithTimeout(context.Background(), 30*time.Second)
-		if err := a.db.InsertIntegrityAuditResults(persistCtx, results); err != nil {
-			a.logger.Error("integrity full audit: FAILED TO PERSIST results — audit ran but paper trail is missing",
-				"org_id", orgID, "proofs_checked", len(proofs), "error", err)
-		}
-		persistCancel()
-
-		totalProofs += len(proofs)
-		for _, r := range results {
-			if !r.Passed {
-				totalFailures++
-			}
-		}
-	}
-
-	a.logger.Info("integrity full audit completed",
-		"orgs_checked", len(orgIDs), "proofs_checked", totalProofs, "failures", totalFailures)
-}
-
-// verifyProofsForOrg checks Merkle roots and chain linkage for the given proofs,
-// returning audit results for every check. Both the sampling and full-sweep
-// audit loops call this.
-func (a *App) verifyProofsForOrg(ctx context.Context, orgID uuid.UUID, proofs []storage.IntegrityProof, sweepType string) []storage.IntegrityAuditResult {
-	now := time.Now().UTC()
-	var results []storage.IntegrityAuditResult
-
-	for i, p := range proofs {
-		// Verify Merkle root.
-		// Prefer snapshotted proof_leaves (survives retention purge and GDPR erasure).
-		// Fall back to re-querying decisions for proofs created before migration 082.
-		hashes, err := a.db.GetProofLeaves(ctx, orgID, p.ID)
-		if err != nil {
-			a.logger.Warn("integrity audit: failed to fetch proof leaves",
-				"org_id", orgID, "proof_id", p.ID, "error", err)
-			results = append(results, storage.IntegrityAuditResult{
-				OrgID: orgID, ProofID: p.ID, CheckType: "merkle_root",
-				Passed: false, SweepType: sweepType,
-				Detail: fmt.Sprintf("failed to fetch proof leaves: %v", err), CheckedAt: now,
-			})
-			continue
-		}
-		if len(hashes) == 0 {
-			// No snapshotted leaves — fall back to decisions table (pre-082 proofs).
-			hashes, err = a.db.GetDecisionHashesForBatch(ctx, orgID, p.BatchStart, p.BatchEnd)
-			if err != nil {
-				a.logger.Warn("integrity audit: failed to fetch hashes for batch",
-					"org_id", orgID, "proof_id", p.ID, "error", err)
-				results = append(results, storage.IntegrityAuditResult{
-					OrgID: orgID, ProofID: p.ID, CheckType: "merkle_root",
-					Passed: false, SweepType: sweepType,
-					Detail: fmt.Sprintf("failed to fetch hashes: %v", err), CheckedAt: now,
-				})
-				continue
-			}
-		}
-
-		ok, err := integrity.VerifyBatchProof(p.RootHash, hashes)
-		if err != nil {
-			a.logger.Error("integrity audit: merkle verification error",
-				"org_id", orgID, "proof_id", p.ID, "error", err)
-			results = append(results, storage.IntegrityAuditResult{
-				OrgID: orgID, ProofID: p.ID, CheckType: "merkle_root",
-				Passed: false, SweepType: sweepType,
-				Detail: fmt.Sprintf("verification error: %v", err), CheckedAt: now,
-			})
-			continue
-		}
-		if !ok {
-			a.logger.Error("INTEGRITY VIOLATION: Merkle root mismatch — stored root does not match recomputed root",
-				"org_id", orgID, "proof_id", p.ID,
-				"stored_root", p.RootHash, "decision_count", p.DecisionCount)
-			a.integrityViolations.Add(ctx, 1, otelmetric.WithAttributes(
-				attribute.String("violation_type", "merkle_root_mismatch"),
-				attribute.String("sweep_type", sweepType),
-			))
-			a.persistViolation(ctx, orgID, p.ID, "merkle_root_mismatch", map[string]any{
-				"stored_root":    p.RootHash,
-				"decision_count": p.DecisionCount,
-				"batch_start":    p.BatchStart,
-				"batch_end":      p.BatchEnd,
-				"leaf_count":     len(hashes),
-			})
-		}
-		results = append(results, storage.IntegrityAuditResult{
-			OrgID: orgID, ProofID: p.ID, CheckType: "merkle_root",
-			Passed: ok, SweepType: sweepType, CheckedAt: now,
-		})
-
-		// Verify chain linkage: this proof's previous_root should match the
-		// next-older proof's root_hash (proofs are newest-first).
-		if i+1 < len(proofs) {
-			older := proofs[i+1]
-			var linkPassed bool
-			var detail string
-			switch {
-			case p.PreviousRoot == nil:
-				a.logger.Warn("integrity audit: proof has nil previous_root but older proof exists — chain may be broken",
-					"org_id", orgID, "proof_id", p.ID, "older_proof_id", older.ID)
-				detail = fmt.Sprintf("nil previous_root but older proof %s exists", older.ID)
-				a.integrityViolations.Add(ctx, 1, otelmetric.WithAttributes(
-					attribute.String("violation_type", "chain_linkage_broken"),
-					attribute.String("sweep_type", sweepType),
-				))
-				a.persistViolation(ctx, orgID, p.ID, "chain_linkage_nil_previous", map[string]any{
-					"older_proof_id": older.ID,
-					"older_root":     older.RootHash,
-				})
-			case *p.PreviousRoot != older.RootHash:
-				a.logger.Error("INTEGRITY VIOLATION: chain linkage broken — previous_root does not match prior proof",
-					"org_id", orgID, "proof_id", p.ID,
-					"expected_previous", older.RootHash, "actual_previous", *p.PreviousRoot)
-				detail = fmt.Sprintf("expected %s, got %s", older.RootHash, *p.PreviousRoot)
-				a.integrityViolations.Add(ctx, 1, otelmetric.WithAttributes(
-					attribute.String("violation_type", "chain_linkage_broken"),
-					attribute.String("sweep_type", sweepType),
-				))
-				a.persistViolation(ctx, orgID, p.ID, "chain_linkage_broken", map[string]any{
-					"expected_previous": older.RootHash,
-					"actual_previous":   *p.PreviousRoot,
-					"older_proof_id":    older.ID,
-				})
-			default:
-				linkPassed = true
-			}
-			results = append(results, storage.IntegrityAuditResult{
-				OrgID: orgID, ProofID: p.ID, CheckType: "chain_linkage",
-				Passed: linkPassed, SweepType: sweepType, Detail: detail, CheckedAt: now,
-			})
-		}
-	}
-
-	return results
-}
-
-// persistViolation writes an integrity violation to the database with retry.
-// This is the durable counterpart to the INTEGRITY VIOLATION log messages —
-// the log is for operators, this record survives log rotation and is queryable
-// for incidents.
-//
-// Because the entire purpose of the violations table is to provide a durable
-// record that survives log rotation, a transient Postgres error (connection
-// blip, disk full) must not silently swallow the violation. We retry up to 3
-// times with exponential backoff before giving up. The violation payload is
-// small and the insert is idempotent (UUID is generated once, before retries).
-func (a *App) persistViolation(ctx context.Context, orgID, proofID uuid.UUID, violationType string, details map[string]any) {
-	v := storage.IntegrityViolation{
-		ID:            uuid.New(),
-		OrgID:         orgID,
-		ProofID:       proofID,
-		ViolationType: violationType,
-		Details:       details,
-		CreatedAt:     time.Now(),
-	}
-
-	const maxRetries = 3
-	backoff := 500 * time.Millisecond
-
-	var lastErr error
-	for attempt := range maxRetries {
-		if err := a.db.CreateIntegrityViolation(ctx, v); err != nil {
-			lastErr = err
-			a.logger.Warn("integrity audit: violation persist attempt failed, will retry",
-				"org_id", orgID, "proof_id", proofID, "violation_type", violationType,
-				"attempt", attempt+1, "max_retries", maxRetries, "error", err)
-
-			select {
-			case <-ctx.Done():
-				a.logger.Error("integrity audit: context cancelled during violation persist retry — violation detected but not durably stored",
-					"org_id", orgID, "proof_id", proofID, "violation_type", violationType, "error", ctx.Err())
-				return
-			case <-time.After(backoff):
-				backoff *= 2
-			}
-			continue
-		}
-		return // success
-	}
-
-	a.logger.Error("integrity audit: exhausted retries for violation persist — violation detected but not durably stored",
-		"org_id", orgID, "proof_id", proofID, "violation_type", violationType,
-		"attempts", maxRetries, "last_error", lastErr)
-}
-
-func (a *App) idempotencyCleanupLoop(ctx context.Context) {
-	a.runLoop(ctx, "idempotencyCleanup", a.cfg.IdempotencyCleanupInterval, func(ctx context.Context) {
-		opCtx, cancel := context.WithTimeout(ctx, 30*time.Second)
-		defer cancel()
-		deleted, err := a.db.CleanupIdempotencyKeys(opCtx, a.cfg.IdempotencyCompletedTTL, a.cfg.IdempotencyAbandonedTTL)
-		if err != nil {
-			a.logger.Warn("idempotency cleanup failed", "error", err)
-			return
-		}
-		if deleted > 0 {
-			a.logger.Info("idempotency cleanup deleted rows", "deleted", deleted)
-		}
-	})
-}
-
-func (a *App) hookCheckCleanupLoop(ctx context.Context) {
-	a.runLoop(ctx, "hookCheckCleanup", 10*time.Minute, func(_ context.Context) {
-		a.srv.Handlers().CleanupHookChecks()
-	})
-}
-
-func (a *App) retentionLoop(ctx context.Context) {
-	if a.cfg.RetentionInterval <= 0 {
-		return
-	}
-	a.runLoop(ctx, "retention", a.cfg.RetentionInterval, func(ctx context.Context) {
-		a.runRetention(ctx)
-	})
-}
-
-func (a *App) claimEmbeddingRetryLoop(ctx context.Context) {
-	if a.cfg.ClaimRetryInterval <= 0 {
-		return
-	}
-	a.runLoop(ctx, "claimEmbeddingRetry", a.cfg.ClaimRetryInterval, func(ctx context.Context) {
-		opCtx, cancel := context.WithTimeout(ctx, a.cfg.ClaimRetryInterval)
-		defer cancel()
-		n, err := a.decisionSvc.RetryFailedClaimEmbeddings(opCtx, 50, 3)
-		if err != nil {
-			a.logger.Warn("claim embedding retry failed", "error", err)
-		} else if n > 0 {
-			a.logger.Info("claim embedding retry complete", "retried", n)
-		}
-	})
-}
-
-// percentileRefreshLoop periodically recomputes signal percentile breakpoints for all orgs
-// and stores them in the in-memory cache. ReScore uses these to normalize citation counts
-// into distribution-aware [0,1] scores instead of using the hardcoded log saturation formula.
-func (a *App) percentileRefreshLoop(ctx context.Context) {
-	if a.cfg.PercentileRefreshInterval <= 0 {
-		return
-	}
-
-	// Refresh immediately on startup to populate the cache before the first search.
-	a.refreshPercentiles(ctx)
-
-	ticker := time.NewTicker(a.cfg.PercentileRefreshInterval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-			a.refreshPercentiles(ctx)
-		}
-	}
-}
-
-func (a *App) refreshPercentiles(ctx context.Context) {
-	opCtx, cancel := context.WithTimeout(ctx, 30*time.Second)
-	defer cancel()
-
-	orgIDs, err := a.db.ListOrganizationIDs(opCtx)
-	if err != nil {
-		a.logger.Warn("percentile refresh: list orgs", "error", err)
-		return
-	}
-
-	refreshed := 0
-	for _, orgID := range orgIDs {
-		bp, err := a.db.GetCitationPercentilesForOrg(opCtx, orgID)
-		if err != nil {
-			a.logger.Warn("percentile refresh: org query failed", "org_id", orgID, "error", err)
-			continue
-		}
-		a.percentileCache.Set(orgID, search.OrgPercentiles{
-			CitationBreakpoints: bp,
-			RefreshedAt:         time.Now(),
-		})
-		refreshed++
-	}
-
-	if refreshed > 0 {
-		a.logger.Debug("percentile refresh complete", "orgs", refreshed)
-	}
-}
-
-// autoResolveLoop periodically runs auto-resolution of conflicts based on org policies.
-func (a *App) autoResolveLoop(ctx context.Context) {
-	if a.cfg.AutoResolveInterval <= 0 {
-		return
-	}
-	a.runLoop(ctx, "autoResolve", a.cfg.AutoResolveInterval, func(ctx context.Context) {
-		if err := a.autoResolver.RunOnce(ctx); err != nil {
-			a.logger.Warn("auto-resolve loop failed", "error", err)
-		}
-	})
-}
-
-// runRetention processes data retention policies for all orgs that have a
-// retention_days set. Each org gets its own deletion_log entry. A global
-// pass also prunes detector-inferred supersedes suggestions older than
-// SupersedesSuggestionTTL — these are operational hints, not user data, and
-// are pruned on a single global TTL regardless of per-org policy.
-func (a *App) runRetention(ctx context.Context) {
-	opCtx, cancel := context.WithTimeout(ctx, a.cfg.RetentionInterval/2)
-	defer cancel()
-
-	a.pruneStaleSupersedesSuggestions(opCtx)
-
-	orgs, err := a.db.GetOrgsWithRetention(opCtx)
-	if err != nil {
-		a.logger.Warn("retention: failed to list orgs with policy", "error", err)
-		return
-	}
-	if len(orgs) == 0 {
-		return
-	}
-
-	for _, org := range orgs {
-		cutoff := time.Now().UTC().AddDate(0, 0, -org.RetentionDays)
-		criteria := map[string]any{"before": cutoff, "retention_days": org.RetentionDays}
-		if len(org.RetentionExcludeTypes) > 0 {
-			criteria["exclude_types"] = org.RetentionExcludeTypes
-		}
-
-		logID, err := a.db.StartDeletionLog(opCtx, org.OrgID, "policy", "", criteria)
-		if err != nil {
-			a.logger.Warn("retention: failed to start deletion log", "org_id", org.OrgID, "error", err)
-			continue
-		}
-
-		counts, err := a.db.BatchDeleteDecisions(opCtx, org.OrgID, cutoff, nil, nil, org.RetentionExcludeTypes, 1000)
-		if err != nil {
-			a.logger.Warn("retention: batch delete failed", "org_id", org.OrgID, "error", err)
-			// Still complete the log even on partial failure so the run is recorded.
-		}
-
-		countMap := map[string]any{
-			"decisions":    counts.Decisions,
-			"alternatives": counts.Alternatives,
-			"evidence":     counts.Evidence,
-			"claims":       counts.Claims,
-			"events":       counts.Events,
-		}
-		if cerr := a.db.CompleteDeletionLog(opCtx, org.OrgID, logID, countMap); cerr != nil {
-			a.logger.Warn("retention: failed to complete deletion log", "org_id", org.OrgID, "error", cerr)
-		}
-
-		if counts.Decisions > 0 || counts.Events > 0 {
-			a.logger.Info("retention: deleted stale records",
-				"org_id", org.OrgID,
-				"decisions", counts.Decisions,
-				"events", counts.Events,
-				"cutoff", cutoff,
-			)
-		}
-	}
-}
-
-// pruneStaleSupersedesSuggestions removes detector-inferred supersedes
-// suggestions that an agent never confirmed. Suggestions confirmed by a
-// re-trace are retired by the migration-106 trigger at confirm time; this
-// global TTL pass cleans up the long tail. Failures are warn-logged but do
-// not abort the retention run — the next tick will retry.
-func (a *App) pruneStaleSupersedesSuggestions(ctx context.Context) {
-	if a.cfg.SupersedesSuggestionTTL <= 0 {
-		return
-	}
-	cutoff := time.Now().UTC().Add(-a.cfg.SupersedesSuggestionTTL)
-	n, err := a.db.DeleteOldSupersedesSuggestions(ctx, cutoff)
-	if err != nil {
-		a.logger.Warn("retention: prune supersedes suggestions failed", "error", err, "cutoff", cutoff)
-		return
-	}
-	if n > 0 {
-		a.logger.Info("retention: pruned stale supersedes suggestions", "count", n, "cutoff", cutoff)
-	}
-}
-
 // ── Adapters (defined here because this file imports both sides) ───────────────
-
-// decisionHookAdapter wraps an akashi.EventHook to satisfy server.DecisionHook.
-// It converts internal model types to public akashi types at the boundary.
-type decisionHookAdapter struct {
-	hook EventHook
-}
-
-func (a *decisionHookAdapter) OnDecisionTraced(ctx context.Context, d model.Decision) error {
-	return a.hook.OnDecisionTraced(ctx, toPublicDecision(d))
-}
-
-func (a *decisionHookAdapter) OnConflictDetected(ctx context.Context, c model.DecisionConflict) error {
-	return a.hook.OnConflictDetected(ctx, toPublicConflict(c))
-}
-
-// externalScorerAdapter wraps an akashi.ConflictScorer to satisfy conflicts.PairwiseScorer.
-type externalScorerAdapter struct {
-	scorer ConflictScorer
-}
-
-func (a *externalScorerAdapter) ScorePair(ctx context.Context, da, db model.Decision) (float32, string, error) {
-	result, err := a.scorer.Score(ctx, toPublicDecision(da), toPublicDecision(db))
-	if err != nil {
-		return 0, "", err
-	}
-	return result.Score, result.Explanation, nil
-}
-
-// searcherAdapter wraps an akashi.Searcher to satisfy search.Searcher.
-// Converts between public SearchFilters/SearchResult and internal model types.
-type searcherAdapter struct {
-	s Searcher
-}
-
-func (a *searcherAdapter) Search(ctx context.Context, orgID uuid.UUID, emb []float32, filters model.QueryFilters, limit int) ([]search.Result, error) {
-	pubFilters := SearchFilters{
-		AgentIDs:      filters.AgentIDs,
-		DecisionType:  filters.DecisionType,
-		ConfidenceMin: filters.ConfidenceMin,
-		SessionID:     filters.SessionID,
-		Tool:          filters.Tool,
-		Model:         filters.Model,
-		Project:       filters.Project,
-	}
-	results, err := a.s.Search(ctx, orgID, emb, pubFilters, limit)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]search.Result, len(results))
-	for i, r := range results {
-		out[i] = search.Result{DecisionID: r.DecisionID, Score: r.Score}
-	}
-	return out, nil
-}
-
-func (a *searcherAdapter) Healthy(ctx context.Context) error {
-	return a.s.Healthy(ctx)
-}
-
-// authHelperImpl implements akashi.AuthHelper using an internal server.RoleMiddlewareFn.
-// Constructed in the route registrar adapter closure; bridges the public interface
-// to the internal RBAC middleware without importing server from enterprise code.
-type authHelperImpl struct {
-	roleFn server.RoleMiddlewareFn
-}
-
-func (a *authHelperImpl) RequireRole(role Role) func(http.Handler) http.Handler {
-	return a.roleFn(model.AgentRole(role))
-}
 
 // ── Type converters ────────────────────────────────────────────────────────────
 
@@ -1563,183 +746,6 @@ func derefOr[T any](ptr *T, fallback T) T {
 }
 
 // ── Helpers (moved from cmd/akashi/main.go) ────────────────────────────────────
-
-func newEmbeddingProvider(cfg config.Config, logger *slog.Logger) embedding.Provider {
-	dims := cfg.EmbeddingDimensions
-
-	switch cfg.EmbeddingProvider {
-	case "openai":
-		if cfg.OpenAIAPIKey == "" {
-			logger.Error("OPENAI_API_KEY required when AKASHI_EMBEDDING_PROVIDER=openai")
-			return embedding.NewNoopProvider(dims)
-		}
-		logger.Info("embedding provider: openai", "model", cfg.EmbeddingModel, "dimensions", dims)
-		p, err := embedding.NewOpenAIProvider(cfg.OpenAIAPIKey.Value(), cfg.EmbeddingModel, dims)
-		if err != nil {
-			logger.Error("openai provider init failed", "error", err)
-			return embedding.NewNoopProvider(dims)
-		}
-		return p
-	case "ollama":
-		logger.Info("embedding provider: ollama", "url", cfg.OllamaURL, "model", cfg.OllamaModel, "dimensions", dims)
-		return embedding.NewOllamaProvider(cfg.OllamaURL, cfg.OllamaModel, dims)
-	case "noop":
-		logger.Info("embedding provider: noop (semantic search disabled)")
-		return embedding.NewNoopProvider(dims)
-	case "auto":
-		fallthrough
-	default:
-		if ollamaReachable(cfg.OllamaURL) {
-			logger.Info("embedding provider: ollama (auto-detected)", "url", cfg.OllamaURL, "model", cfg.OllamaModel, "dimensions", dims)
-			return embedding.NewOllamaProvider(cfg.OllamaURL, cfg.OllamaModel, dims)
-		}
-		if cfg.OpenAIAPIKey != "" {
-			logger.Info("embedding provider: openai (auto-detected)", "model", cfg.EmbeddingModel, "dimensions", dims)
-			p, err := embedding.NewOpenAIProvider(cfg.OpenAIAPIKey.Value(), cfg.EmbeddingModel, dims)
-			if err != nil {
-				logger.Error("openai provider init failed", "error", err)
-				return embedding.NewNoopProvider(dims)
-			}
-			return p
-		}
-		logger.Warn("no embedding provider available, using noop (semantic search disabled)")
-		return embedding.NewNoopProvider(dims)
-	}
-}
-
-func newConflictValidator(cfg config.Config, logger *slog.Logger) conflicts.Validator {
-	if cfg.ConflictLLMModel != "" {
-		logger.Info("conflict validator: ollama", "model", cfg.ConflictLLMModel, "url", cfg.OllamaURL, "num_threads", cfg.ConflictLLMThreads)
-		return conflicts.NewOllamaValidator(cfg.OllamaURL, cfg.ConflictLLMModel, cfg.ConflictLLMThreads)
-	}
-	if cfg.OpenAIAPIKey != "" {
-		logger.Info("conflict validator: openai", "model", cfg.ConflictOpenAIModel, "timeout", cfg.ConflictLLMTimeout)
-		return conflicts.NewOpenAIValidator(cfg.OpenAIAPIKey.Value(), cfg.ConflictOpenAIModel,
-			conflicts.WithRequestTimeout(cfg.ConflictLLMTimeout))
-	}
-	logger.Info("conflict validator: noop (no LLM configured, embedding-only conflicts)")
-	return conflicts.NoopValidator{}
-}
-
-// newClaimExtractor creates an LLM-backed claim extractor when configured.
-// Returns nil when LLM claim extraction is disabled or no LLM is available.
-func newClaimExtractor(cfg config.Config, logger *slog.Logger) conflicts.ClaimExtractor {
-	if !cfg.ClaimExtractionLLM {
-		return nil
-	}
-	if cfg.ConflictLLMModel != "" {
-		logger.Info("claim extractor: ollama", "model", cfg.ConflictLLMModel, "url", cfg.OllamaURL)
-		return conflicts.NewOllamaExtractor(cfg.OllamaURL, cfg.ConflictLLMModel, cfg.ConflictLLMThreads)
-	}
-	if cfg.OpenAIAPIKey != "" {
-		logger.Info("claim extractor: openai (gpt-4o-mini)")
-		return conflicts.NewOpenAIExtractor(cfg.OpenAIAPIKey.Value(), "gpt-4o-mini")
-	}
-	logger.Warn("claim extraction LLM requested but no LLM configured, using regex fallback")
-	return nil
-}
-
-func ollamaReachable(baseURL string) bool {
-	c, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-	req, err := http.NewRequestWithContext(c, http.MethodGet, baseURL+"/api/tags", nil)
-	if err != nil {
-		return false
-	}
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return false
-	}
-	_, _ = io.Copy(io.Discard, resp.Body)
-	_ = resp.Body.Close()
-	return resp.StatusCode == http.StatusOK
-}
-
-func buildIntegrityProofs(ctx context.Context, db *storage.DB, logger *slog.Logger) {
-	orgIDs, err := db.ListOrganizationIDs(ctx)
-	if err != nil {
-		logger.Warn("integrity proof: list orgs failed", "error", err)
-		return
-	}
-
-	now := time.Now().UTC()
-
-	for _, orgID := range orgIDs {
-		latest, err := db.GetLatestIntegrityProof(ctx, orgID)
-		if err != nil {
-			logger.Warn("integrity proof: get latest failed", "error", err, "org_id", orgID)
-			continue
-		}
-
-		batchStart := time.Time{}
-		var previousRoot *string
-		if latest != nil {
-			batchStart = latest.BatchEnd
-			previousRoot = &latest.RootHash
-		}
-
-		hashes, err := db.GetDecisionHashesForBatch(ctx, orgID, batchStart, now)
-		if err != nil {
-			logger.Warn("integrity proof: get hashes failed", "error", err, "org_id", orgID)
-			continue
-		}
-		if len(hashes) == 0 {
-			continue
-		}
-
-		root, err := integrity.BuildMerkleRoot(hashes)
-		if err != nil {
-			logger.Warn("integrity proof: merkle root construction failed", "error", err, "org_id", orgID)
-			continue
-		}
-
-		proofID := uuid.New()
-		proof := storage.IntegrityProof{
-			ID:            proofID,
-			OrgID:         orgID,
-			BatchStart:    batchStart,
-			BatchEnd:      now,
-			DecisionCount: len(hashes),
-			RootHash:      root,
-			PreviousRoot:  previousRoot,
-			CreatedAt:     now,
-		}
-
-		if err := db.CreateIntegrityProof(ctx, proof); err != nil {
-			logger.Warn("integrity proof: create failed", "error", err, "org_id", orgID)
-			continue
-		}
-
-		// Snapshot the leaf hashes so verification survives retention purge
-		// and GDPR erasure (which delete or mutate the decisions table).
-		if err := db.CreateProofLeaves(ctx, proofID, orgID, hashes); err != nil {
-			logger.Warn("integrity proof: save leaves failed", "error", err, "org_id", orgID, "proof_id", proofID)
-			// The proof itself was created; leaves can be backfilled later.
-			// Verification will fall back to GetDecisionHashesForBatch.
-		}
-
-		logger.Info("integrity proof created",
-			"org_id", orgID,
-			"decisions", len(hashes),
-			"root_hash", root[:16]+"...",
-		)
-	}
-}
-
-func contextWithOptionalTimeout(parent context.Context, timeout time.Duration) (context.Context, context.CancelFunc) {
-	if timeout <= 0 {
-		return context.WithCancel(parent)
-	}
-	return context.WithTimeout(parent, timeout)
-}
-
-func conflictRefreshTimeout(interval time.Duration) time.Duration {
-	const max = 15 * time.Second
-	if interval < max {
-		return interval
-	}
-	return max
-}
 
 // ctx is a no-op helper so that New(opts ...) can pass a background context to
 // telemetry.Init without adding a context parameter to the public API.
